@@ -1,6 +1,6 @@
 {
   "app": {
-    "name": "Test",
+    "name": "react-starter",
     "version": "0.0.1",
     "build": "1"
   },
@@ -31,67 +31,66 @@
   },
   "dependencies": {
     "fastclick": "1.0.6",
-    "history": "2.1.2",
+    "history": "4.5.1",
     "immutable": "3.8.1",
-    "react": "15.3.2",
-    "react-dom": "15.3.2",
-    "react-redux": "4.4.5",
-    "react-router": "2.8.1",
+    "react": "15.4.2",
+    "react-dom": "15.4.2",
+    "react-redux": "5.0.2",
+    "react-router": "3.0.1",
     "redux": "3.6.0",
-    "redux-thunk": "2.1.0",
-    "whatwg-fetch": "1.0.0"
+    "redux-thunk": "2.2.0",
+    "whatwg-fetch": "2.0.1"
   },
   "devDependencies": {
-    "babel-cli": "6.16.0",
-    "babel-core": "6.17.0",
-    "babel-eslint": "7.0.0",
-    "babel-jest": "16.0.0",
-    "babel-loader": "6.2.5",
+    "babel-cli": "6.18.0",
+    "babel-core": "6.21.0",
+    "babel-eslint": "7.1.1",
+    "babel-jest": "18.0.0",
+    "babel-loader": "6.2.10",
     "babel-plugin-add-module-exports": "0.2.1",
     "babel-plugin-transform-react-constant-elements": "6.9.1",
     "babel-plugin-transform-react-inline-elements": "6.8.0",
     "babel-plugin-transform-react-remove-prop-types": "0.2.10",
     "babel-plugin-transform-runtime": "6.15.0",
-    "babel-polyfill": "6.16.0",
-    "babel-preset-es2015": "6.16.0",
+    "babel-polyfill": "6.20.0",
+    "babel-preset-es2015": "6.18.0",
     "babel-preset-es2015-webpack": "6.4.3",
     "babel-preset-react": "6.16.0",
     "babel-preset-react-hmre": "1.1.1",
     "babel-preset-stage-1": "6.16.0",
-    "babel-register": "6.16.3",
-    "babel-runtime": "6.11.6",
-    "browser-sync": "2.17.3",
+    "babel-register": "6.18.0",
+    "babel-runtime": "6.20.0",
+    "browser-sync": "2.18.6",
     "connect-history-api-fallback": "1.3.0",
-    "css-loader": "0.25.0",
-    "electron": "1.3.3",
-    "electron-packager": "7.5.1",
-    "enzyme": "2.4.1",
-    "eslint": "3.8.0",
-    "eslint-config-airbnb": "12.0.0",
-    "eslint-import-resolver-webpack": "0.6.0",
-    "eslint-plugin-import": "1.16.0",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.4.1",
+    "css-loader": "0.26.1",
+    "electron": "1.4.14",
+    "electron-packager": "8.5.0",
+    "eslint": "3.13.1",
+    "eslint-config-airbnb": "14.0.0",
+    "eslint-import-resolver-webpack": "0.8.0",
+    "eslint-plugin-import": "2.2.0",
+    "eslint-plugin-jsx-a11y": "3.0.2",
+    "eslint-plugin-react": "6.9.0",
     "express": "4.14.0",
     "file-loader": "0.9.0",
     "html-loader": "0.4.4",
-    "jest": "16.0.1",
+    "jest": "18.1.0",
     "json-loader": "0.5.4",
-    "postcss-cssnext": "2.8.0",
+    "postcss-cssnext": "2.9.0",
     "postcss-focus": "1.0.0",
-    "postcss-loader": "0.13.0",
-    "postcss-reporter": "1.4.1",
-    "react-addons-test-utils": "15.3.2",
-    "react-hot-loader": "3.0.0-beta.6",
-    "react-test-renderer": "15.3.2",
+    "postcss-loader": "1.2.2",
+    "postcss-reporter": "3.0.0",
+    "react-addons-test-utils": "15.4.2",
+    "react-hot-loader": "1.3.1",
+    "react-test-renderer": "15.4.2",
     "sinon": "1.17.6",
     "style-loader": "0.13.1",
-    "stylelint": "7.4.2",
-    "stylelint-config-standard": "13.0.2",
+    "stylelint": "7.7.1",
+    "stylelint-config-standard": "15.0.1",
     "url-loader": "0.5.7",
-    "webpack": "1.13.2",
-    "webpack-dev-middleware": "1.8.4",
-    "webpack-hot-middleware": "2.13.0",
+    "webpack": "1.14.0",
+    "webpack-dev-middleware": "1.9.0",
+    "webpack-hot-middleware": "2.15.0",
     "webpack-target-electron-renderer": "0.4.0"
   },
   "jest": {
