@@ -1,5 +1,1 @@
-import userActions from './user';
-
-module.exports = {
-    ...userActions
-};
+export default {};
